ENH Improve efficiency of chi2 by converting the input to float first

+++ sklearn/feature_selection/_univariate_selection.py
@@ -210,7 +210,7 @@ def chi2(X, y):
 
     # XXX: we might want to do some of the following in logspace instead for
     # numerical stability.
-    X = check_array(X, accept_sparse="csr")
+    X = check_array(X, accept_sparse="csr", dtype=np.float64)
     if np.any((X.data if issparse(X) else X) < 0):
         raise ValueError("Input X must be non-negative.")
